Renamed MainActivity -> EnterActivity Refactored EnterActivity
