+++ app/src/main/AndroidManifest.xml
@@ -38,7 +38,7 @@
             android:label="@string/forgot_password_activity"
             android:theme="@style/Theme.Chessgo" />
         <activity
-            android:name=".frontend.MainActivity"
+            android:name=".frontend.registration.enterApp.EnterActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Chessgo">
