+++ app/src/main/java/com/example/chessgo/frontend/registration/enterApp/EnterActivityCompose.kt
@@ -1,0 +1,50 @@
+package com.example.chessgo.frontend.registration.enterApp
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun GreetingScreen(enterActivity: EnterActivity,) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            text = "Welcome!",
+            style = TextStyle(fontSize = 30.sp),
+            modifier = Modifier.padding(bottom = 32.dp)
+        )
+
+        Button(
+            onClick = { enterActivity.goToSignIn() },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+        ) {
+            Text(text = "Log In")
+        }
+
+        Button(
+            onClick = { enterActivity.goToSignUp() },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+        ) {
+            Text(text = "Register")
+        }
+    }
+}
